+++ frontend/src/components/prompts/Prompts.vue
@@ -36,6 +36,7 @@ import FileList from "./FileList.vue";
 import SaveBeforeExit from "./SaveBeforeExit.vue";
 import CopyPasteConfirm from "./CopyPasteConfirm.vue";
 import CloseWithActiveUploads from "./CloseWithActiveUploads.vue";
+import Generic from "./Generic.vue";
 import { state, getters, mutations } from "@/store"; // Import your custom store
 
 export default {
@@ -68,6 +69,7 @@ export default {
     SaveBeforeExit,
     CopyPasteConfirm,
     CloseWithActiveUploads,
+    generic: Generic,
   },
   data() {
     return {
@@ -104,6 +106,9 @@ export default {
             break;
           case "replace":
             prompt.showConfirm(event);
+            break;
+          case "generic":
+            prompt.submit();
             break;
           case "CopyPasteConfirm":
             prompt.confirm();